HSM-343: Modify comments for key exchange kdf_input to add hello randoms
On further study of the TLS spec, the client/server random values are
different when calculating the master_secret and the key_block.  When
calculating the master secret we need to use ClientHello.random and
ServerHello.random (concatenated in that order), and when calculating
the key block we need to use SecurityParameters.server_random and
SecurityParameters.client_random (concatenated in that order).
Update the comments for kdf_input (size & data) to reflect this, also
specifying the order they are provided so that they can be used by the
code without copies.

Signed-off-by: Michael Whitfield <michael.whitfield@nxp.com>