Makefile: Fix install to clear host-user-contaminated
warnings [HSM-483]

WARNING: imx-seco-libs-git-r0 do_package_qa: QA Issue:
imx-seco-libs: /usr/lib/hsm_lib.a is owned by uid 26292, which is the same as the user running bitbake. This may be due to host contamination
imx-seco-libs: /usr/lib/she_lib.a is owned by uid 26292, which is the same as the user running bitbake. This may be due to host contamination
imx-seco-libs: /usr/lib/seco_nvm_manager.a is owned by uid 26292, which is the same as the user running bitbake.
This may be due to host contamination [host-user-contaminated]

Signed-off-by: Cristinel Panfir <cristinel.panfir@nxp.com>