Fix install with usrmerge
The logic allowing the toolchain to set the library install location
for multilib does not work when usrmerge is enabled, resulting in the
following incorrect install:
```
  /usr/usr/lib/libele_hsm.so.1
  /usr/usr/lib/libele_hsm.so
  /usr/usr/lib/libele_nvm.so
  /usr/usr/lib/libele_hsm.so.1.0
  /usr/usr/lib/libele_nvm.so.1.0
  /usr/usr/lib/libele_nvm.so.1
```

The problem is the use of the toolchain variable `base_libdir`, which
does take the value /usr/lib when usrmerge is enabled.

Fix the problem by using the toolchain variable `libdir` instead.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>